requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord==2.3.1
discord.py==2.3.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numerize==0.12
numpy==1.25.0
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
riotwatcher==3.2.5
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
